Remove beta guard for appName LogRecordContext feature
