+++ dev/com.ibm.ws.collector.manager/bnd.bnd
@@ -37,14 +37,9 @@ Import-Package: com.ibm.ejs.ras, \
                 com.ibm.wsspi.collector.manager, \
                 com.ibm.ws.collector.manager.buffer, \
                 com.ibm.websphere.logging, \
-                com.ibm.websphere.logging.hpel, \
                 com.ibm.tools.attach, \
                 org.osgi.service.cm, \
-                com.ibm.wsspi.logging, \
-                com.ibm.ws.threadContext;resolution:=dynamic, \
-                com.ibm.ws.runtime.metadata;resolution:=dynamic, \
-                com.ibm.websphere.csi;resolution:=dynamic, \
-                com.ibm.ws.kernel.productinfo
+                com.ibm.wsspi.logging
 
 Service-Component:\
     com.ibm.ws.collector.manager; \
@@ -77,10 +72,7 @@ Include-Resource: \
 	com.ibm.ws.transport.http;version=latest,\
 	com.ibm.ws.logging;version=latest,\
 	com.ibm.ws.logging.core;version=latest,\
-	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-	com.ibm.ws.container.service;version=latest,\
-	com.ibm.ws.container.service.compat;version=latest,\
-	com.ibm.ws.kernel.boot.core;version=latest
+	com.ibm.ws.org.osgi.annotation.versioning;version=latest
 
 -testpath: \
 	../build.sharedResources/lib/junit/old/junit.jar;version=file, \
