--- dev/com.ibm.ws.logging.json_fat/publish/servers/com.ibm.ws.logging.json.AppNameExtensionServer/jvm.options
@@ -1,1 +1,0 @@
--Dcom.ibm.ws.beta.edition=true
